ml/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
pdfminer.six
python-docx
transformers
torch
python-dotenv
requests
